requirements.txt
streamlit
together
pandas
python-dotenv
